Move isE2EBuild to separate file
Some of our imports can error in Storybook so moving this util function
out makes sense.
