+++ node-src/lib/e2e.ts
@@ -1,6 +1,6 @@
 import { AGENTS, getCliCommand, Runner } from '@antfu/ni';
 
-import { Context, Options } from '../types';
+import { Context } from '../types';
 import missingDependency from '../ui/messages/errors/missingDependency';
 import { failed } from '../ui/tasks/build';
 import { exitCodes, setExitCode } from './setExitCode';
@@ -65,15 +65,4 @@ export async function getE2EBuildCommand(
 
     throw err;
   }
-}
-
-/**
- * Determine if the build is an E2E build.
- *
- * @param options Parsed options when executing the CLI (usually from the context).
- *
- * @returns true if the build is an E2E build.
- */
-export function isE2EBuild(options: Options) {
-  return options.playwright || options.cypress;
 }
